moved landing form to pages folder
